main: switch to using sync.Map
Instead of locking the whole map, lock per-key by using a sync.Map which
already does everything for us. This helps with the cases where there
are a lot of writers and only a few readers. Add a new field
`numsamples' which tracks the number of samples since sync.Map does not
have a Len() method.

Signed-off-by: Giedrius Statkevičius <giedrius.statkevicius@adform.com>